fix prompt submission + alpha relate + loop sound
Prompt Submission: In our eagerness to remove "unusued" functions, ScrubWrite() was simplified to "ScrubBuffer = []". Consequently, scrubbing became impossible as there was no handler to consume scrub requests into actual scrub actions. This change has been reverted.

AlphaRelate: Although the function does work, there are a ton of off-by-one errors throughout the code. I have revised the code to not put _ in the required prompts and "Enter" in the viruses... these did not make sense and were impossible to solve

Sound Looping: The music was not looping. By setting the CHANNEL.play(SOUND, -1) we can specify the number of times to loop in arg[1]. For QED, this should be 0, but for INTRO and MUSIC it should be -1 (infinate)