merged stacked prompt + added backspace
Stacked Prompt: On the MoreServers branch, I devised a system where the user types the prompt on top of the problem, resulting in a sort of "fill in the blank" system

Backspace: The user can now backspace, but only if they can spend 25 energy doing so. 25 energy is also what it costs to hit enter, so hitting enter to just "give up" is still preferable for typos of more than one character

Alphabet error: Alphabet.RandomSafe() would fallback to CAPITAL letters, which were unreachable. these have been revised to their lowercase counterparts