fallback warning screen
if for whatever reason font initalization fails, enter GameActive -1, a screen that displays error information. In this case, it tells the user to install the font manually or proceed with bad graphics. In the event of an error in the font loading, the traceback is dumped to console but execution continues