switch to fixed-prompt system
instead of users typing a random series of letters, there is now a fixed list of prompts (located in `/assets/prompts.txt`), which is randomly polled each time a prompt is needed. these prompts are simpler but are much longer to make up for that. care as also been taken to ensure that at least 1 prompt uses every letter of the English alphabet.